+++ .idea/dictionaries/epool.xml
@@ -4,6 +4,7 @@
       <w>gson</w>
       <w>nearbooks</w>
       <w>nearsoft</w>
+      <w>zxing</w>
     </words>
   </dictionary>
 </component>
