+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/nearbooks.iml" filepath="$PROJECT_DIR$/nearbooks.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Nearbooks.iml" filepath="$PROJECT_DIR$/Nearbooks.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
