+++ app/src/main/java/com/nearsoft/nearbooks/di/components/GoogleApiClientComponent.java
@@ -1,9 +1,9 @@
 package com.nearsoft.nearbooks.di.components;
 
-import com.nearsoft.nearbooks.activities.GoogleApiClientBaseActivity;
 import com.nearsoft.nearbooks.di.modules.BaseActivityModule;
 import com.nearsoft.nearbooks.di.modules.GoogleApiClientModule;
 import com.nearsoft.nearbooks.di.scopes.PerActivity;
+import com.nearsoft.nearbooks.view.activities.GoogleApiClientBaseActivity;
 
 import dagger.Component;
