+++ app/src/main/java/com/nearsoft/nearbooks/view/activities/BaseActivity.java
@@ -1,4 +1,4 @@
-package com.nearsoft.nearbooks.activities;
+package com.nearsoft.nearbooks.view.activities;
 
 import android.databinding.DataBindingUtil;
 import android.databinding.ViewDataBinding;
