+++ app/src/main/java/com/nearsoft/nearbooks/view/realm/RealmRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.nearsoft.nearbooks.adapters.realm;
+package com.nearsoft.nearbooks.view.realm;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
