+++ app/src/main/res/layout/activity_book_detail.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="book"
-            type="com.nearsoft.nearbooks.models.viewmodels.BookViewModel" />
+            type="com.nearsoft.nearbooks.view.models.BookViewModel" />
     </data>
 
     <android.support.design.widget.CoordinatorLayout
