+++ app/src/main/res/layout/fragment_book_detail.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="book"
-            type="com.nearsoft.nearbooks.models.viewmodels.BookViewModel" />
+            type="com.nearsoft.nearbooks.view.models.BookViewModel" />
     </data>
 
     <LinearLayout
