+++ app/src/main/res/layout/fragment_library.xml
@@ -14,7 +14,7 @@
             android:visibility="gone" />
 
         <TextView
-            android:id="@android:id/empty"
+            android:id="@+id/empty"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
